add: getSurvey

+++ src/main/java/com/nexters/winepick/survey/repository/SurveyRepository.java
@@ -3,5 +3,5 @@ package com.nexters.winepick.survey.repository;
 import com.nexters.winepick.survey.domain.Survey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SurveyRepository extends JpaRepository<Survey, Long> {
+public interface SurveyRepository extends JpaRepository<Survey, Integer> {
 }
